Changed two values.
I changed the width of #username to auto. I didn't see the point of having it to 50%, and it just look neater :3

And about #headContainer. Since the height was to 10%, if, let's say your screen is squashed, the username will overflow over the container and it looks wrong.

These might be bad ideas, sorry ^^" I'm not much of a web dev, haha.