border instead of box-shadow
I saw that #statsUpper had a box-shadow on it, to draw that horizontal line which rules off the column labels. I think it'd be better to use the border property. Better you a more supported property rather than something like box-shadow, in essence of cross-browser compatibility.

statsMiddle seems to be commented out so I didn't apply the border to lines 129-132.

Instead of deleting it I just added another block; I would've deleted the 129 block, though I'll leave that up to you.

(:

And yea, I get a bit worked up when it comes to cross-browser compatability *.*